+++ packages/calcite-components/src/components/list-item-group/list-item-group.e2e.ts
@@ -1,5 +1,6 @@
-import { html } from "../../../support/formatting";
+import { describe } from "vitest";
 import { hidden, renders, disabled, defaults, themed } from "../../tests/commonTests";
+import { html } from "../../../support/formatting";
 import { CSS } from "./resources";
 
 describe("calcite-list-item-group", () => {